prevent duplicate cache entries
Without preserving the entryId in the AfterAccessNotification, in line 97 the EntryId is always equal to 0 which always causes a new record to be written to the database.

Because the class is used on a scoped level within Startup.cs and is constructed specific to a user, general member functions should not cause a wipe of all cache entries for all users.  Clearing for the currently scoped user also allows for clearing of expired tokens when a new access token cannot be silently obtained.