additional info for protecting against http redirection
Added additional comment information for protecting against undesired redirection to http resources when hosting provider on a reverse proxy that proxies and rewrites https traffic to http. We experienced this with a hacker-security web-application-firewall service, where even when communication persisted through HTTPS tunnels, the URL was still rewritten from https to http. This then resulted in initial form post back being sent to http as well as the post-authorization code received action redirecting to http instead of https. Adding replaces of http:// to https:// on context.ProtocolMessage.RedirectUri and context.Properties.RedirectUri resolved the issue.